add styling
