+++ packages/ddex-frontend/src/components/admin/Table.module.css
@@ -1,0 +1,38 @@
+.styledTable {
+  width: 100%;
+  border-collapse: collapse;
+  margin-bottom: 25px;
+  margin-top: 8px;
+  font-size: 0.9em;
+  font-family: sans-serif;
+  min-width: 400px;
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.15);
+  border-radius: 4px;
+  overflow: hidden;
+}
+
+.styledTable thead tr {
+  background-color: var(--harmony-secondary);
+  color: var(--harmony-white);
+  text-align: left;
+}
+
+.styledTable th,
+.styledTable td {
+  padding: 12px 15px;
+}
+
+.styledTable tbody tr:nth-of-type(even) {
+  /* background-color: #f3f3f3; */
+  background-color: var(--harmony-background);
+}
+
+.styledTable tbody tr:last-of-type {
+  border-bottom: 2px solid var(--harmony-secondary);
+}
+
+.styledTable tbody tr.activeRow {
+  font-weight: bold;
+  color: var(--harmony-secondary);
+}
+
