+++ packages/ddex-frontend/src/main.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 
 import ReactDOM from 'react-dom/client'
 
-import AppWithProviders from './components/AppWithProviders'
-import { EnvVarsProvider } from './providers/EnvVarsProvider'
+import AppWithProviders from 'components/AppWithProviders'
+import { EnvVarsProvider } from 'providers/EnvVarsProvider'
 
 import '@audius/harmony/dist/harmony.css'
