+++ packages/ddex-frontend/src/pages/DDEX/DDEX.module.css
@@ -1,3 +1,6 @@
+.page {
+  height: 100vh;
+}
 .fileDrop {
   display: flex;
   justify-content: center;
@@ -7,7 +10,7 @@
   transition-property: color, background-color, border-color, text-decoration-color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter;
   transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
   transition-duration: 150ms;
-  background-color: var(--static-white);
+  background-color: var(--harmony-white);
   border-width: 2px;
   border-style: dashed;
   border-radius: 0.375rem;
@@ -57,4 +60,10 @@
 
 .successText {
   color: var(--harmony-green);
+}
+
+.adminBadge {
+  border: 2px solid;
+  border-radius: 4px;
+  padding: 8px;
 }
