+++ packages/ddex-frontend/src/pages/DDEX/DDEX.tsx
@@ -1,6 +1,13 @@
 import { useState, DragEvent } from 'react'
 
-import { Button, ButtonType, ButtonSize, Box, Flex } from '@audius/harmony'
+import {
+  Text,
+  Button,
+  ButtonType,
+  ButtonSize,
+  Box,
+  Flex
+} from '@audius/harmony'
 import type {
   DecodedUserToken,
   AudiusSdk
@@ -8,8 +15,8 @@ import type {
 import { useQueryClient } from '@tanstack/react-query'
 import cn from 'classnames'
 
-import Releases from 'components/Releases'
-import Uploads from 'components/Uploads'
+import Releases from 'components/admin/Releases'
+import Uploads from 'components/admin/Uploads'
 import { useAudiusSdk } from 'providers/AudiusSdkProvider'
 
 import styles from './DDEX.module.css'
@@ -31,17 +38,27 @@ const validXmlFile = (file: File) => {
 
 const ManageAudiusAccount = ({
   currentUser,
+  isAdmin,
   onChangeUser,
   oauthError
 }: {
   currentUser: DecodedUserToken
+  isAdmin: boolean
   onChangeUser: () => void
   oauthError: string | null
 }) => {
   return (
     <Flex justifyContent='space-between' alignItems='center'>
-      <div>{`Logged in as @${currentUser.handle}`}</div>
-      <Button variant="secondary" onClick={onChangeUser}>
+      <Flex gap='m' alignItems='center'>
+        <Text
+          variant='body'
+          color='default'
+        >{`Logged in as @${currentUser.handle}`}</Text>
+        <Text variant='label' color='subdued' className={styles.adminBadge}>
+          {isAdmin ? 'ADMIN' : ''}
+        </Text>
+      </Flex>
+      <Button variant={ButtonType.SECONDARY} onClick={onChangeUser}>
         Switch users
       </Button>
       {oauthError && <div className='text-red-600'>{oauthError}</div>}
@@ -149,73 +166,84 @@ const XmlImporter = ({
     return <div className='text-red-500'>{'Error loading XML importer'}</div>
   } else {
     return (
-      <>
-        <label
-          className={cn(
-            styles.fileDrop,
-            { [styles.fileDropBorderDragging]: isDragging },
-            { [styles.fileDropBorder]: !isDragging }
-          )}
-          onDragEnter={handleDragIn}
-          onDragLeave={handleDragOut}
-          onDragOver={handleDragOver}
-          onDrop={handleDrop}
-        >
-          <span className={styles.fileDropTextContainer}>
-            <svg
-              xmlns='http://www.w3.org/2000/svg'
-              className={styles.fileDropIcon}
-              fill='none'
-              viewBox='0 0 24 24'
-              stroke='currentColor'
-              strokeWidth='2'
-            >
-              <path
-                strokeLinecap='round'
-                strokeLinejoin='round'
-                d='M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12'
-              />
-            </svg>
-            <span>
-              {'Drop files to upload, or '}
-              <span className={styles.fileDropBrowseText}>browse</span>
-            </span>
-          </span>
-          <input
-            type='file'
-            name='file_upload'
-            accept='text/xml,application/xml'
-            className={styles.fileDropChooseFile}
-            onChange={(e) => handleFileChange(e.target.files![0])}
-          />
-        </label>
-        {selectedFile && (
-          <>
-            <div>Selected file:</div>
-            <Flex gap='xs'>
-              <div>{selectedFile.name}</div>
-              <Button
-                variant="destructive"
-                size="small"
-                onClick={clearSelection}
+      <Box borderRadius='s' shadow='near' p='xl' backgroundColor='white'>
+        <Flex direction='column' gap='l'>
+          <Text variant='heading' color='heading'>
+            Upload a DDEX delivery
+          </Text>
+          <label
+            className={cn(
+              styles.fileDrop,
+              { [styles.fileDropBorderDragging]: isDragging },
+              { [styles.fileDropBorder]: !isDragging }
+            )}
+            onDragEnter={handleDragIn}
+            onDragLeave={handleDragOut}
+            onDragOver={handleDragOver}
+            onDrop={handleDrop}
+          >
+            <span className={styles.fileDropTextContainer}>
+              <svg
+                xmlns='http://www.w3.org/2000/svg'
+                className={styles.fileDropIcon}
+                fill='none'
+                viewBox='0 0 24 24'
+                stroke='currentColor'
+                strokeWidth='2'
               >
-                x
+                <path
+                  strokeLinecap='round'
+                  strokeLinejoin='round'
+                  d='M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12'
+                />
+              </svg>
+              <Text variant='body' color='default'>
+                {'Drop files to upload, or '}
+                <span className={styles.fileDropBrowseText}>browse</span>
+              </Text>
+            </span>
+            <input
+              type='file'
+              name='file_upload'
+              accept='text/xml,application/xml'
+              className={styles.fileDropChooseFile}
+              onChange={(e) => handleFileChange(e.target.files![0])}
+            />
+          </label>
+          {selectedFile && (
+            <>
+              <Text variant='body' color='default'>
+                Selected file:
+              </Text>
+              <Flex gap='s' alignItems='center'>
+                <Button
+                  variant={ButtonType.DESTRUCTIVE}
+                  size={ButtonSize.SMALL}
+                  onClick={clearSelection}
+                >
+                  x
+                </Button>
+                <Text variant='body' color='default'>
+                  {selectedFile.name}
+                </Text>
+              </Flex>
+              <Button onClick={handleUpload} disabled={isUploading}>
+                {isUploading ? 'Uploading...' : 'Upload'}
               </Button>
-            </Flex>
-            <Button onClick={handleUpload} disabled={isUploading}>
-              {isUploading ? 'Uploading...' : 'Upload'}
-            </Button>
-            {uploadError && (
-              <div className={styles.errorText}>
-                {`Error uploading file: ${uploadError}`}
-              </div>
-            )}
-            {uploadSucceeded && (
-              <div className={styles.successText}>Upload success!</div>
-            )}
-          </>
-        )}
-      </>
+              {uploadError && (
+                <Text variant='body' className={styles.errorText}>
+                  {`Error uploading file: ${uploadError}`}
+                </Text>
+              )}
+              {uploadSucceeded && (
+                <Text variant='body' className={styles.successText}>
+                  Upload success!
+                </Text>
+              )}
+            </>
+          )}
+        </Flex>
+      </Box>
     )
   }
 }
@@ -240,12 +268,13 @@ const Ddex = () => {
   // }
 
   return (
-    <Box m='xl'>
+    <Box>
       <Flex gap='xs' direction='column'>
         {!audiusSdk ? (
           'loading...'
         ) : !currentUser ? (
           <Flex
+            p='xl'
             gap='m'
             direction='column'
             justifyContent='center'
@@ -255,31 +284,43 @@ const Ddex = () => {
             {oauthError && <div className={styles.errorText}>{oauthError}</div>}
           </Flex>
         ) : (
-          <Flex gap='xl' direction='column'>
-            <ManageAudiusAccount
-              currentUser={currentUser}
-              onChangeUser={handleOauth}
-              oauthError={oauthError}
-            />
-            <XmlImporter audiusSdk={audiusSdk} />
-            {/* <ManageAudiusAccount
-              currentUser={currentUser || fakeUser}
-              isAdmin={true}
-              onChangeUser={handleOauth}
-              oauthError={oauthError}
-            />
-            <XmlImporter audiusSdk={audiusSdk} /> */}
+          <Flex direction='column'>
+            <Box p='xl'>
+              <ManageAudiusAccount
+                currentUser={currentUser}
+                isAdmin={isAdmin}
+                onChangeUser={handleOauth}
+                oauthError={oauthError}
+              />
+            </Box>
+            <Box
+              p='xl'
+              backgroundColor='surface2'
+              borderTop='strong'
+              className={styles.page}
+            >
+              <Flex gap='xl' direction='column'>
+                <XmlImporter audiusSdk={audiusSdk} />
+                {/* <ManageAudiusAccount
+                  currentUser={currentUser || fakeUser}
+                  isAdmin={true}
+                  onChangeUser={handleOauth}
+                  oauthError={oauthError}
+                />
+                <XmlImporter audiusSdk={audiusSdk} /> */}
 
-            {isAdmin && (
-              <>
-                <Flex direction='column' gap='s'>
-                  <Uploads />
-                </Flex>
-                <Flex direction='column' gap='s'>
-                  <Releases />
-                </Flex>
-              </>
-            )}
+                {isAdmin && (
+                  <>
+                    <Flex direction='column' gap='s'>
+                      <Uploads />
+                    </Flex>
+                    <Flex direction='column' gap='s'>
+                      <Releases />
+                    </Flex>
+                  </>
+                )}
+              </Flex>
+            </Box>
           </Flex>
         )}
       </Flex>
